refactor(scheduler): dedupe buildFailureCountPhrase into a shared util
Slack, Google Chat, and MS Teams each had a byte-for-byte identical
per-kind failure count builder; extract it to utils/partialFailureUtils
so the three clients import one implementation instead of drifting.
